Working version with registered IdP
This is the first version working with the idp registration feature
It bypass the client id check from 3.5.2 ( https://fedidcg.github.io/FedCM/#idp-api-id-assertion-endpoint )
But I feel it is good enough for a demo. I just need to clean up and refactor
a little and then will merge this version to main.