A cyber-frankenstein that works with inrupt authn lib
Return a location with a authz code through fedcm token
So make CSS truly work with FedCM
However the code is very gross, because for this POC
we took the front end code for OIDC auth and put it
in our backend, and try to automate the OIDC flow bypassing
the login and consent.
Legacy from front-end code, we therefore fetch the backend
from the backend, which is not elegant but digging into
oidc-provider librairy we couldn't found the right function
to call, so we fallback on calling them through their API
endpoint. But we will refactor that once we have a better
undertanding of the oidc-provider lib