\o/ removed last fetch to get session_id
We still needed a fetch to get a session id
We realised we could totally remove the use
of session since we do all in one shot without
the need for user interactin. The catch here
is that oidc-provider use webId value for its
`accountId` which is confusing since CSS also
uses `accountId` values, but which are totally
differents from webIds.